+++ modules/core/localizer.js
@@ -3,7 +3,7 @@ import { escape } from 'es-toolkit/compat';
 
 import { fileFetcher } from './file_fetcher';
 import { utilDetect } from '../util/detect';
-import { utilStringQs } from '../util';
+import { utilExpandLocaleCode, utilStringQs } from '../util';
 import { utilArrayUniq } from '../util/array';
 import { presetsCdnUrl } from '../../config/id.js';
 
@@ -48,6 +48,7 @@ export function coreLocalizer() {
     let _localeCode = 'en-US';
     // `_localeCodes` must contain `_localeCode` first, optionally followed by fallbacks
     let _localeCodes = ['en-US', 'en'];
+    let _expandedLocaleCodes = utilExpandLocaleCode(_localeCode);
     let _languageCode = 'en';
     let _textDirection = 'ltr';
     let _usesMetric = false;
@@ -57,6 +58,7 @@ export function coreLocalizer() {
     // getters for the current locale parameters
     localizer.localeCode = () => _localeCode;
     localizer.localeCodes = () => _localeCodes;
+    localizer.expandedLocaleCodes = () => _expandedLocaleCodes;
     localizer.languageCode = () => _languageCode;
     localizer.textDirection = () => _textDirection;
     localizer.usesMetric = () => _usesMetric;
@@ -114,6 +116,7 @@ export function coreLocalizer() {
 
                 _localeCodes = localizer.localesToUseFrom(_dataLocales);
                 _localeCode = _localeCodes[0];   // Run iD in the highest-priority locale; the rest are fallbacks
+                _expandedLocaleCodes = utilExpandLocaleCode(_localeCode);
 
                 let loadStringsPromises = [];