+++ modules/util/index.ts
@@ -1,16 +1,6 @@
 export { utilAesEncrypt } from './aes';
 export { utilAesDecrypt } from './aes';
-
-export { utilArrayChunk } from './array';
-export { utilArrayDifference } from './array';
-export { utilArrayFlatten } from './array';
-export { utilArrayGroupBy } from './array';
-export { utilArrayIdentical } from './array';
-export { utilArrayIntersection } from './array';
-export { utilArrayUnion } from './array';
-export { utilArrayUniq } from './array';
-export { utilArrayUniqBy } from './array';
-
+export * from './array';
 export { utilCleanTags } from './clean_tags';
 export { utilCombinedTags } from './util';
 export { utilDeepMemberSelector } from './util';
@@ -31,6 +21,7 @@ export { utilGetSetValue } from './get_set_value';
 export { utilHashcode } from './util';
 export { utilHighlightEntities } from './util';
 export { utilKeybinding } from './keybinding';
+export * from './localizer';
 export { utilNoAuto } from './util';
 export { utilObjectOmit, utilCheckTagDictionary, stringifyProperties } from './object';
 export { utilCompareIDs } from './util';